[HAPI-95] Move all Schemas to collection folder (#112)
* [HAPI-95] Move all Schemas to collection folder

## [HAPI-95] Move all Schemas to collection folder

* [HAPI-95] [HAPI-95] Move all Schemas to collection folder

Removing the files
