+++ src/collections/Class/_globals.class.js
