+++ src/collections/Class/safe_values.class.js
