+++ src/collections/Class/users.class.js
