+++ src/collections/_globals.js
