+++ src/collections/map/Logstamp.js
