+++ src/collections/map/index.js
