--- src/schemas/class/index.js
@@ -1,13 +1,0 @@
-const GlobalClass = require('./GlobalClass');
-const auth_buckets = require('./AuthBucketsClass');
-const logs = require('./LogsClass');
-const users = require('./UsersClass');
-const safe_values = require('./safe_values.class');
-
-module.exports = {
-    GlobalClass,
-    auth_buckets,
-    logs,
-    users,
-    safe_values
-};
