--- src/schemas/events/index.js
@@ -1,9 +1,0 @@
-const users = require('./users_events');
-const auth_buckets = require('./auth_buckets_events');
-const safe_values = require('./safe_values.event');
-
-module.exports = {
-    users,
-    auth_buckets,
-    safe_values
-};
