--- src/schemas/queries/index.js
@@ -1,7 +1,0 @@
-const users = require('./users_query');
-const safe_values = require('./safe_values.query');
-
-module.exports = {
-    users,
-    safe_values
-};
